Added generic locking primative.
This change adds one new ioctl command TENSTORRENT_IOCTL_LOCK_CTL which
as input takes a flag describing the operation and lock index to perfom
the operation on. The command returns a value containing the result of
the command.

The intentions is that the provides a set of generic locks (currently
64) which can be used to associate resources with a device handle. If a
device handle is closed then all of the resources associated with the
handle are automatically released.

The operations are
TENSTORRENT_LOCK_CTL_ACQUIRE: returns 1 if the lock was acquired; 0 if
it was not
TENSTORRENT_LOCK_CTL_RELEASE: returns 1 if the lock was released; 0 if
it was not
TENSTORRENT_LOCK_CTL_TEST: return 0b10 if the lock is held by any handle
to the device; Or'd with 0b01 if the calling handle holds the lock.