Workaround for ETH core sharing
Problem: ETH core request and response queues can not be shared between
processes without synchronization, and KMD uses an ETH core in its hwmon
code path for N300 devices.  Although KMD exposes a locking API,
userspace is free to interact with ETH cores without using this API.

Workaround: track whether processes with an open device have used the
locking API.  If all such processes have used it, assume they are using
it correctly.  If any such process has not used it, assume that we can
not access an ETH core safely and return -EBUSY in the code path for
reading telemetry from the remote chip on an N300 board.