fix(blackhole): Wait for ARC readiness before sending messages
After a device reset, the driver would attempt to communicate with the
ARC firmware before it was ready to process messages.

This resulted in the failure of the two messages sent by the driver,
which are responsible for setting the initial ASIC state and the
watchdog timer.

To fix this, `send_arc_message` now polls a scratch register for a
readiness flag that is set by the ARC firmware upon successful
initialization.

This change guarantees that the driver does not attempt to send messages
until the ARC firmware is ready.