Always build for all kernel versions and hotload module
Usually, `dkms build` builds for the most recent kernel version. However,
when installing the package also triggers a kernel update, the module is
built for the to-be-loaded new kernel, and thus the tenstorrent module
is unavailable on the running kernel. To fix this, we explicitly build
and load the module for the running kernel before doing the normal dkms
install command.

Signed-off-by: June Knauth (tt) <jknauth@tenstorrent.com>